♻️ refactor : 유즈케이스 및 서비스 폴더 위치️ 수정
