--- src/main/java/com/seeat/server/domain/review/application/ReviewService.java
@@ -1,7 +1,0 @@
-package com.seeat.server.domain.review.application;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class ReviewService implements ReviewUseCase{
-}
