--- src/main/java/com/seeat/server/domain/review/application/ReviewUseCase.java
@@ -1,4 +1,0 @@
-package com.seeat.server.domain.review.application;
-
-public interface ReviewUseCase {
-}
