+++ src/main/java/com/seeat/server/domain/review/application/usecase/BookmarkUseCase.java
@@ -1,0 +1,4 @@
+package com.seeat.server.domain.review.application.usecase;
+
+public interface BookmarkUseCase {
+}
