+++ src/main/java/com/seeat/server/domain/review/application/usecase/ReviewUseCase.java
@@ -1,0 +1,12 @@
+package com.seeat.server.domain.review.application.usecase;
+
+
+/**
+ * [리뷰 인터페이스]
+ * - 로그인한 유저만 좌석 후기를 작성 및 열람할 수 있습니다.
+ * - 로그인한 유저만 좌석 클릭 시 해당 좌석의 후기 및 사진 열람 가능
+ * -
+ */
+public interface ReviewUseCase {
+
+}
